feat: Single revision data returning endpoint & translation filtering (#2284)
I have added 
- endpoint which returns data for activity in paginated form
- filter option, that allows to filter translations by provided revision
id (only keys mentioned in revision are returned)

**Why this?**
- When user executes operation modifying many columns, there is no way
in the system where you can check what specifically was changed.
- When operations are executed, users are notified in slack and can
usually can get to the platform to key edition view by clicking a button
in the Slack message. This works fine when single key is affected by the
operation. However, we need to provide similar path when larger amount
data is imported so user is redirected to the Activity view and from
there, they're able to go to translations view.

This is blocking the Slack integration, so we need simple solution
before we implement the ultimate activity grouping and notification
system.

@stepan662 can you please add frontend for the Activity view? This
should be pretty the same as it is now, we just need to add pagination
@ZuzanaOdstrcilova can you create a UI proposal of how it should look in
the translations view
- IMO, we should display this filter information super clearly, since
this is not a standard filter, which is selectable from filter menu
- We can show - the date / time of the activity, the autor, and the
activity type (e.g., key creation, translations update, import, etc)
 - Let's discuss it Monday.
@huglx Can you please sync with Stepan, so we redirect users from Slack
to the activity when import is executed?

Thanks!