feat: Add translation update time to the stats endpoint (#3018)
We now expose a timestamp of the last modification as explain in #3008 
I planned to write tests, however I wasn't sure where and what kind of
tests you prefer. I would like to have some guidance, and then I can add
tests as well.

I tested this manually by updating translations for a language, and then
I expected the timestamp to change.
In the video, you can see that I was checking the network tab in the
devtools which also query the stats endpoint.

The interesting part is the new timestmap which changes from 
`1743062340802` which is `Thu Mar 27 2025 07:59:00.802` to
`1743063177609`  which is `Thu Mar 27 2025 08:12:57.609`


https://github.com/user-attachments/assets/40cf8990-c8da-4751-a79e-f526b5e3e54c



<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Language statistics now include a “Last Translation Update” timestamp,
giving you clear insight into when translations were last refreshed.
  
- **Bug Fixes**
- Improved data retrieval for language statistics to accurately reflect
the last modified date of translations.

- **Tests**
- Enhanced test coverage confirms the new timestamp is correctly
reflected in language statistics responses.
- Additional validation ensures updates to language statistics reflect
the most recent translation update time.
- New assertions added to verify the presence of the last translation
update timestamp in responses.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Stefan Schaller <stefan.schaller@tapped.dev>