fix: tests in TranslationSuggestionControllerStreamingTest were randomly failing with ConcurrentModificationException;
Spring is trying to read response headers as part of handling stream
write calls and something is trying to modify them at the same time.
Issue might be specific to tests, as the code causing the exception
is specific to mocking streaming responses, but I was unable to find
any info about multithreading safety guaranties by spring when it
comes to streaming responses so it might be safer to avoid writing
the results from multiple threads asynchronously anyway.