fix: import with meta without creating new keys  (#3000) (#3113)
When `shouldImportKey` is called in the
`FileProcessorContext.processTranslations()` method, Hibernate does the
flush before executing a select query. At this flush, it discovers that
`ImportKey` has `KeyMeta`, which is newly created and needs to be
flushed, however, there is no cascading.

So cadacding would be a good solution. It is also the first one which is
recommended in many articles, e.g.
https://www.baeldung.com/hibernate-unsaved-transient-instance-error, as
well as cascading is usually (sure, not always) applied in OneToOne
relations:
https://vladmihalcea.com/the-best-way-to-map-a-onetoone-relationship-with-jpa-and-hibernate/.
But when I started applying cascading, I ended up modifying too many
entities, which led to many unpredictable problems

So I think this solution of saving a `KeyMeta` alongside the `ImportKey`
is a good compromise here.

Probably the test can be moved to a more suitable place. (I'm just
getting acquainted with the code base). It can easily be replicated also
with the `strings.xml` file mentioned in [the
comment](https://github.com/tolgee/tolgee-platform/issues/3000#issuecomment-2865753203)

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Tests**
- Added a new test case to verify import behavior when creating new keys
is disabled.
- Updated existing tests to include the new "createNewKeys" setting
during import.
- **New Features**
- Improved handling of key metadata during import, ensuring author
information is set correctly for comments and code references.
- Enhanced import process to prepare and save key metadata individually
for each key.
- **Bug Fixes**
- Fixed import of PO files with single key metadata when creating new
keys is disabled.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: dmitrii.bocharov <bdshadow@gmail.com>