fix: select chunks from db, which are not locked
Having just LockOptions.SKIP_LOCKED actually does nothing. In the postgres docs https://www.postgresql.org/docs/current/sql-select.html
it can be found that "SKIP LOCKED" is always used with "FOR { UPDATE | NO KEY UPDATE |...". Moreover, hibernate is smart enough and didn't
even add SKIP LOCKED, when there was not LockModeType.PESSIMISTIC_WRITE (instruction for "FOR UPDATE"). It can be seen by turning on the
hibernate's "showSql=true"

Without this, there were rare cases when LOCKED chunks were returned from database and sent processing again. During processing in the
`BatchJobActionService.handleItem` it was found out that row is locked, then coroutine is finished and `BatchJobConcurrentLauncher.onJobCompleted`
removed the coroutine job from `runningJobs`. Now cancelling is coming - doesn't find it in the `runningJobs` and the coroutine is no cancelled.

Probably, it also led to conflicts when running on multiple k8s pods - same chunks could be selected simultaneously and sent for processing.
These changes don't fully solve it, because it still can happen, when this transaction in one pod is finished and in another one started. This
will be fixed next (probably by introducing a new chunk status, "NEW", which will be updated in this transaction to "PENDING")

Setting `LockModeType.PESSIMISTIC_WRITE` also required to make it Transactional, because rows are locked only within transaction

Signed-off-by: dmitrii.bocharov <bdshadow@gmail.com>